fix sonarqube issue
