+++ app/screening/screening-row.tsx
@@ -18,6 +18,7 @@ import { deleteScreening } from "@/commons/api-calls.common";
 import { toast } from "react-toastify";
 import { TResponseError } from "@/types/response.type";
 import { screeningKeys } from "./query-key-factories";
+import { onError } from "@/commons/mutation-on-error.common";
 
 const ScreeningRow = ({
   screening,
@@ -44,25 +45,7 @@ const ScreeningRow = ({
           refetchType: "active",
         });
       },
-      onError(error) {
-        const serverResponse = JSON.parse(
-          error.message.replace("Error: ", ""),
-        ) as TResponseError;
-        let errorMessage = "An unknown error has occurred";
-
-        // assigning `serverResponse.detail.message` logic to a variable
-        // leads to the error: property `message` does not exist on type `string`
-        if (
-          typeof serverResponse.detail === "object" &&
-          "message" in serverResponse.detail
-        ) {
-          errorMessage = JSON.stringify(serverResponse.detail.message);
-        } else if (typeof serverResponse.detail === "string") {
-          errorMessage = serverResponse.detail;
-        }
-
-        toast.error(errorMessage);
-      },
+      onError,
     });
   };
